feat: moves features to its own package
