--- pkg/kfapp/ossm/cleanup.go
@@ -1,14 +1,0 @@
-package ossm
-
-import (
-	"github.com/hashicorp/go-multierror"
-)
-
-func (o *OssmInstaller) CleanupResources() error {
-	var cleanupErrors *multierror.Error
-	for _, feature := range o.features {
-		cleanupErrors = multierror.Append(cleanupErrors, feature.Cleanup())
-	}
-
-	return cleanupErrors.ErrorOrNil()
-}
