+++ pkg/kfapp/ossm/feature/cert.go
@@ -1,4 +1,4 @@
-package ossm
+package feature
 
 import (
 	"bytes"
