+++ pkg/kfapp/ossm/feature/envoy_secrets.go
@@ -1,4 +1,4 @@
-package ossm
+package feature
 
 import (
 	"bytes"
