+++ pkg/kfapp/ossm/feature/feature_raw_resources.go
@@ -11,7 +11,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package ossm
+package feature
 
 import (
 	"context"
