+++ pkg/kfapp/ossm/feature/feature_suite_test.go
@@ -1,0 +1,14 @@
+package feature_test
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+)
+
+func TestOssmFeatures(t *testing.T) {
+	RegisterFailHandler(Fail)
+	// for integration tests see tests/integration directory
+	RunSpecs(t, "Openshift Service Mesh features unit tests")
+}
