+++ pkg/kfapp/ossm/feature/types.go
@@ -1,0 +1,18 @@
+package feature
+
+import (
+	"strings"
+)
+
+type oAuth struct {
+	AuthzEndpoint,
+	TokenEndpoint,
+	Route,
+	Port,
+	ClientSecret,
+	Hmac string
+}
+
+func ReplaceChar(s string, oldChar, newChar string) string {
+	return strings.ReplaceAll(s, oldChar, newChar)
+}
