+++ pkg/kfapp/ossm/ossm_installer.go
@@ -1,26 +1,21 @@
 package ossm
 
 import (
-	"context"
 	"embed"
 	"fmt"
 	"github.com/hashicorp/go-multierror"
 	kfapisv3 "github.com/opendatahub-io/opendatahub-operator/apis"
 	kftypesv3 "github.com/opendatahub-io/opendatahub-operator/apis/apps"
+	"github.com/opendatahub-io/opendatahub-operator/pkg/kfapp/ossm/feature"
 	"github.com/opendatahub-io/opendatahub-operator/pkg/kfconfig"
 	"github.com/opendatahub-io/opendatahub-operator/pkg/kfconfig/ossmplugin"
 	"github.com/pkg/errors"
-	k8serrors "k8s.io/apimachinery/pkg/api/errors"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
-	"k8s.io/apimachinery/pkg/labels"
-	"k8s.io/apimachinery/pkg/runtime/schema"
+	"io/fs"
 	"k8s.io/client-go/rest"
-	"net/url"
+	"os"
 	"path"
 	"path/filepath"
 	ctrlLog "sigs.k8s.io/controller-runtime/pkg/log"
-	"strconv"
 )
 
 const (
@@ -38,7 +33,7 @@ type OssmInstaller struct {
 	*kfconfig.KfConfig
 	pluginSpec *ossmplugin.OssmPluginSpec
 	config     *rest.Config
-	features   []*Feature
+	features   []*feature.Feature
 }
 
 func NewOssmInstaller(kfConfig *kfconfig.KfConfig, restConfig *rest.Config) *OssmInstaller {
@@ -97,286 +92,98 @@ func (o *OssmInstaller) enableFeatures() error {
 		return internalError(err)
 	}
 
-	var rootDir = filepath.Join(baseOutputDir, o.Namespace, o.Name)
+	var rootDir = filepath.Join(feature.BaseOutputDir, o.Namespace, o.Name)
 	if copyFsErr := copyEmbeddedFS(embeddedFiles, "templates", rootDir); copyFsErr != nil {
 		return internalError(errors.WithStack(copyFsErr))
 	}
 
-	if feature, err := EnableFeature("control-plane-oauth").
+	if oauth, err := feature.CreateFeature("control-plane-oauth").
 		For(o.pluginSpec).
 		WithConfig(o.config).
 		FromPaths(
-			path.Join(rootDir, ControlPlaneDir, "base"),
-			path.Join(rootDir, ControlPlaneDir, "oauth"),
-			path.Join(rootDir, ControlPlaneDir, "filters"),
+			path.Join(rootDir, feature.ControlPlaneDir, "base"),
+			path.Join(rootDir, feature.ControlPlaneDir, "oauth"),
+			path.Join(rootDir, feature.ControlPlaneDir, "filters"),
 		).
 		AdditionalResources(
-			func(feature *Feature) error {
-				if feature.spec.Mesh.Certificate.Generate {
-					meta := metav1.ObjectMeta{
-						Name:      feature.spec.Mesh.Certificate.Name,
-						Namespace: feature.spec.Mesh.Namespace,
-						OwnerReferences: []metav1.OwnerReference{
-							feature.tracker.ToOwnerReference(),
-						},
-					}
-
-					cert, err := generateSelfSignedCertificateAsSecret(feature.data.Domain, meta)
-					if err != nil {
-						return internalError(err)
-					}
-
-					if err != nil {
-						return errors.WithStack(err)
-					}
-
-					_, err = feature.clientset.CoreV1().
-						Secrets(feature.spec.Mesh.Namespace).
-						Create(context.TODO(), cert, metav1.CreateOptions{})
-					if err != nil && !k8serrors.IsAlreadyExists(err) {
-						return errors.WithStack(err)
-					}
-				}
-
-				return nil
-			},
-			func(feature *Feature) error {
-				objectMeta := metav1.ObjectMeta{
-					Name:      feature.spec.AppNamespace + "-oauth2-tokens",
-					Namespace: feature.spec.Mesh.Namespace,
-					OwnerReferences: []metav1.OwnerReference{
-						feature.tracker.ToOwnerReference(),
-					},
-				}
-
-				envoySecret, err := createEnvoySecret(feature.data.OAuth, objectMeta)
-				if err != nil {
-					return internalError(err)
-				}
-
-				_, err = feature.clientset.CoreV1().
-					Secrets(objectMeta.Namespace).
-					Create(context.TODO(), envoySecret, metav1.CreateOptions{})
-				if err != nil && !k8serrors.IsAlreadyExists(err) {
-					return errors.WithStack(err)
-				}
-
-				return nil
-			},
+			feature.GenerateSelfSignedCertificate,
+			feature.GenerateEnvoySecrets,
 		).
-		WithData(loadData).
+		WithData(feature.LoadClusterData).
 		Preconditions(
-			checkIfCRDIsInstalled("operator.authorino.kuadrant.io", "v1beta1", "authorinos"),
-			ensureServiceMeshInstalled,
+			feature.EnsureCRDIsInstalled("operator.authorino.kuadrant.io", "v1beta1", "authorinos"),
+			feature.EnsureServiceMeshInstalled,
 		).
 		OnDelete(
-			removeOAuthClient,
-			removeTokenVolumes,
+			feature.RemoveOAuthClient,
+			feature.RemoveTokenVolumes,
 		).Load(); err != nil {
 		return nil
 	} else {
-		o.features = append(o.features, feature)
+		o.features = append(o.features, oauth)
 	}
 
-	if feature, err := EnableFeature("shared-config-maps").
+	if cfMaps, err := feature.CreateFeature("shared-config-maps").
 		For(o.pluginSpec).
 		WithConfig(o.config).
-		AdditionalResources(
-			func(feature *Feature) error {
-				if err := feature.createConfigMap("service-mesh-refs",
-					map[string]string{
-						"CONTROL_PLANE_NAME": feature.spec.Mesh.Name,
-						"MESH_NAMESPACE":     feature.spec.Mesh.Namespace,
-					}); err != nil {
-					return internalError(err)
-				}
-
-				if err := feature.createConfigMap("auth-refs",
-					map[string]string{
-						"AUTHORINO_LABEL": feature.spec.Auth.Authorino.Label,
-					}); err != nil {
-					return internalError(err)
-				}
-
-				return nil
-			},
-		).Load(); err != nil {
-		return err
-	} else {
-		o.features = append(o.features, feature)
-	}
-
-	if feature, err := EnableFeature("enable-service-mesh").
-		For(o.pluginSpec).
-		WithConfig(o.config).
-		FromPaths(
-			path.Join(rootDir, ControlPlaneDir, "smm.tmpl"),
-			path.Join(rootDir, ControlPlaneDir, "namespace.patch.tmpl"),
-		).
-		WithData(loadData).
+		AdditionalResources(feature.CreateConfigMaps).
 		Load(); err != nil {
 		return err
 	} else {
-		o.features = append(o.features, feature)
+		o.features = append(o.features, cfMaps)
 	}
 
-	if feature, err := EnableFeature("enable-service-mesh-for-dashboard").
-		For(o.pluginSpec).
-		WithConfig(o.config).
-		AdditionalResources(
-			func(feature *Feature) error {
-				gvr := schema.GroupVersionResource{
-					Group:    "opendatahub.io",
-					Version:  "v1alpha",
-					Resource: "odhdashboardconfigs",
-				}
-
-				configs, err := feature.dynamicClient.Resource(gvr).List(context.Background(), metav1.ListOptions{})
-				if err != nil {
-					return err
-				}
-
-				if len(configs.Items) == 0 {
-					log.Info("No odhdashboardconfig found in namespace, doing nothing")
-					return nil
-				}
-
-				// Assuming there is only one odhdashboardconfig in the namespace, patching the first one
-				config := configs.Items[0]
-				if config.Object["spec"] == nil {
-					config.Object["spec"] = map[string]interface{}{}
-				}
-				spec := config.Object["spec"].(map[string]interface{})
-				if spec["dashboardConfig"] == nil {
-					spec["dashboardConfig"] = map[string]interface{}{}
-				}
-				dashboardConfig := spec["dashboardConfig"].(map[string]interface{})
-				dashboardConfig["disableServiceMesh"] = false
-
-				_, err = feature.dynamicClient.Resource(gvr).
-					Namespace(feature.spec.AppNamespace).
-					Update(context.Background(), &config, metav1.UpdateOptions{})
-				if err != nil {
-					log.Error(err, "Failed to update odhdashboardconfig")
-					return err
-				}
-
-				log.Info("Successfully patched odhdashboardconfig")
-				return nil
-
-			},
-		).Load(); err != nil {
-		return err
-	} else {
-		o.features = append(o.features, feature)
-	}
-
-	if feature, err := EnableFeature("migrate-data-science-projects").
-		For(o.pluginSpec).
-		WithConfig(o.config).
-		AdditionalResources(
-			func(feature *Feature) error {
-				selector := labels.SelectorFromSet(labels.Set{"opendatahub.io/dashboard": "true"})
-
-				namespaceClient := feature.clientset.
-					CoreV1().
-					Namespaces()
-
-				namespaces, err := namespaceClient.List(context.TODO(), metav1.ListOptions{LabelSelector: selector.String()})
-				if err != nil {
-					return fmt.Errorf("failed to get namespaces: %v", err)
-				}
-
-				var result *multierror.Error
-
-				for _, namespace := range namespaces.Items {
-					annotations := namespace.GetAnnotations()
-					if annotations == nil {
-						annotations = map[string]string{}
-					}
-					annotations["opendatahub.io/service-mesh"] = "true"
-					namespace.SetAnnotations(annotations)
-
-					if _, err := namespaceClient.Update(context.TODO(), &namespace, metav1.UpdateOptions{}); err != nil {
-						result = multierror.Append(result, err)
-					}
-				}
-
-				return result.ErrorOrNil()
-			},
-		).Load(); err != nil {
-		return err
-	} else {
-		o.features = append(o.features, feature)
-	}
-
-	if feature, err := EnableFeature("setup-external-authorizaion").
+	if serviceMesh, err := feature.CreateFeature("enable-service-mesh").
 		For(o.pluginSpec).
 		WithConfig(o.config).
 		FromPaths(
-			path.Join(rootDir, AuthDir, "namespace.tmpl"),
-			path.Join(rootDir, AuthDir, "auth-smm.tmpl"),
-			path.Join(rootDir, AuthDir, "base"),
-			path.Join(rootDir, AuthDir, "rbac"),
-			path.Join(rootDir, AuthDir, "mesh-authz-ext-provider.patch.tmpl"),
+			path.Join(rootDir, feature.ControlPlaneDir, "smm.tmpl"),
+			path.Join(rootDir, feature.ControlPlaneDir, "namespace.patch.tmpl"),
 		).
-		WithData(loadData).
-		OnDelete(
-			func(feature *Feature) error {
-				ossmAuthzProvider := fmt.Sprintf("%s-odh-auth-provider", feature.spec.AppNamespace)
-
-				gvr := schema.GroupVersionResource{
-					Group:    "maistra.io",
-					Version:  "v2",
-					Resource: "servicemeshcontrolplanes",
-				}
-
-				mesh := feature.spec.Mesh
-
-				smcp, err := feature.dynamicClient.Resource(gvr).
-					Namespace(mesh.Namespace).
-					Get(context.Background(), mesh.Name, metav1.GetOptions{})
-				if err != nil {
-					return err
-				}
-
-				extensionProviders, found, err := unstructured.NestedSlice(smcp.Object, "spec", "techPreview", "meshConfig", "extensionProviders")
-				if err != nil {
-					return err
-				}
-				if !found {
-					log.Info("no extension providers found", "smcp", mesh.Name, "istio-ns", mesh.Namespace)
-					return nil
-				}
-
-				for i, v := range extensionProviders {
-					extensionProvider, ok := v.(map[string]interface{})
-					if !ok {
-						fmt.Println("Unexpected type for extensionProvider")
-						continue
-					}
-
-					if extensionProvider["name"] == ossmAuthzProvider {
-						extensionProviders = append(extensionProviders[:i], extensionProviders[i+1:]...)
-						err = unstructured.SetNestedSlice(smcp.Object, extensionProviders, "spec", "techPreview", "meshConfig", "extensionProviders")
-						if err != nil {
-							return err
-						}
-						break
-					}
-				}
-
-				_, err = feature.dynamicClient.Resource(gvr).
-					Namespace(mesh.Namespace).
-					Update(context.Background(), smcp, metav1.UpdateOptions{})
-
-				return err
-
-			},
-		).Load(); err != nil {
+		WithData(feature.LoadClusterData).
+		Load(); err != nil {
 		return err
 	} else {
-		o.features = append(o.features, feature)
+		o.features = append(o.features, serviceMesh)
+	}
+
+	if dashboard, err := feature.CreateFeature("enable-service-mesh-for-dashboard").
+		For(o.pluginSpec).
+		WithConfig(o.config).
+		AdditionalResources(feature.EnableServiceMeshInDashboard).
+		Load(); err != nil {
+		return err
+	} else {
+		o.features = append(o.features, dashboard)
+	}
+
+	if dataScienceProjects, err := feature.CreateFeature("migrate-data-science-projects").
+		For(o.pluginSpec).
+		WithConfig(o.config).
+		// TODO this is not creating any resource - it is updating it
+		AdditionalResources(feature.MigrateDataScienceProjects).
+		Load(); err != nil {
+		return err
+	} else {
+		o.features = append(o.features, dataScienceProjects)
+	}
+
+	if extAuthz, err := feature.CreateFeature("setup-external-authorization").
+		For(o.pluginSpec).
+		WithConfig(o.config).
+		FromPaths(
+			path.Join(rootDir, feature.AuthDir, "namespace.tmpl"),
+			path.Join(rootDir, feature.AuthDir, "auth-smm.tmpl"),
+			path.Join(rootDir, feature.AuthDir, "base"),
+			path.Join(rootDir, feature.AuthDir, "rbac"),
+			path.Join(rootDir, feature.AuthDir, "mesh-authz-ext-provider.patch.tmpl"),
+		).
+		WithData(feature.LoadClusterData).
+		OnDelete(feature.RemoveExtensionProvider).
+		Load(); err != nil {
+		return err
+	} else {
+		o.features = append(o.features, extAuthz)
 	}
 
 	return nil
@@ -385,45 +192,21 @@ func (o *OssmInstaller) enableFeatures() error {
 func (o *OssmInstaller) Generate(_ kftypesv3.ResourceEnum) error {
 	var applyErrors *multierror.Error
 
-	for _, feature := range o.features {
-		err := feature.Apply()
+	for _, f := range o.features {
+		err := f.Apply()
 		applyErrors = multierror.Append(applyErrors, err)
 	}
 
 	return applyErrors.ErrorOrNil()
 }
 
-// ExtractHostNameAndPort strips given URL in string from http(s):// prefix and subsequent path,
-// returning host name and port if defined (otherwise defaults to 443).
-//
-// This is useful when getting value from http headers (such as origin).
-// If given string does not start with http(s) prefix it will be returned as is.
-func ExtractHostNameAndPort(s string) (string, string, error) {
-	u, err := url.Parse(s)
-	if err != nil {
-		return "", "", err
+func (o *OssmInstaller) CleanupResources() error {
+	var cleanupErrors *multierror.Error
+	for _, feature := range o.features {
+		cleanupErrors = multierror.Append(cleanupErrors, feature.Cleanup())
 	}
 
-	if u.Scheme != "http" && u.Scheme != "https" {
-		return s, "", nil
-	}
-
-	hostname := u.Hostname()
-
-	port := "443" // default for https
-	if u.Scheme == "http" {
-		port = "80"
-	}
-
-	if u.Port() != "" {
-		port = u.Port()
-		_, err := strconv.Atoi(port)
-		if err != nil {
-			return "", "", errors.New("invalid port number: " + port)
-		}
-	}
-
-	return hostname, port, nil
+	return cleanupErrors.ErrorOrNil()
 }
 
 func internalError(err error) error {
@@ -431,4 +214,33 @@ func internalError(err error) error {
 		Code:    int(kfapisv3.INTERNAL_ERROR),
 		Message: fmt.Sprintf("%+v", err),
 	}
+}
+
+// In order to process the templates, we need to create a tmp directory
+// to store the files. This is because embedded files are read only.
+// copyEmbeddedFS ensures that files embedded using go:embed are populated
+// to dest directory
+func copyEmbeddedFS(fsys fs.FS, root, dest string) error {
+	return fs.WalkDir(fsys, root, func(path string, d fs.DirEntry, err error) error {
+		if err != nil {
+			return err
+		}
+
+		destPath := filepath.Join(dest, path)
+		if d.IsDir() {
+			if err := os.MkdirAll(destPath, 0755); err != nil {
+				return err
+			}
+		} else {
+			data, err := fs.ReadFile(fsys, path)
+			if err != nil {
+				return err
+			}
+			if err := os.WriteFile(destPath, data, 0644); err != nil {
+				return err
+			}
+		}
+
+		return nil
+	})
 }
